requirements.txt
aws-cdk-lib==2.117.0
constructs>=10.0.0,<11.0.0
aws-cdk.lambda-layer-kubectl-v28==2.1.0
cdk8s-plus-27==2.7.70
cdk-nag==2.27.9